Flyttet tekster til intl Renaming av filer
