+++ packages/fakta-soknadsperioder/i18n/nb_NO.json
@@ -1,1 +1,21 @@
-{}
+{
+  "Soknadsperioder.Rad.Vedtakshistorikk": "Vedtakshistorikk",
+  "Soknadsperioder.Rad.PerioderTilBehandling": "Perioder til behandling",
+  "Soknadsperioder.Rad.SøknadOmNyPeriode": "Søknad om ny periode",
+  "Soknadsperioder.Rad.EndringerFraSøker": "Endringer fra søker",
+  "Soknadsperioder.Rad.EndringerPgaAnnenPart": "Endringer pga. annen part",
+  "Soknadsperioder.Rad.NyInntektsmelding": "Ny inntektsmelding",
+  "Soknadsperioder.Rad.BerørtPeriode": "Berørt periode",
+  "Soknadsperioder.Rad.TrukketKrav": "Trukket krav",
+  "Soknadsperioder.Rad.GRegulering": "G-regulering",
+  "Soknadsperioder.Rad.ManuellRevurdering": "Manuell revurdering",
+  "Soknadsperioder.Skala.SkalaForVisning": "Velg skala for visning",
+  "Soknadsperioder.Skala.6mnd": "6 mnd",
+  "Soknadsperioder.Skala.1år": "1 år",
+  "Soknadsperioder.Skala.3år": "3 år",
+  "Soknadsperioder.Status.InnvilgetPeriode": "Innvilget periode",
+  "Soknadsperioder.Status.AvslåttPeriode": "Avslått periode",
+  "Soknadsperioder.Status.TilBehandling": "Til behandling",
+  "Soknadsperioder.SøknadMedDato": "Søknad {dato}",
+  "Soknadsperioder.Søknadsperioder": "Søknadsperioder"
+}
